선택 시간 서버통신할 때 반영 (#120)
* 이전 TimePicker 삭제

* 선택 시간 서버통신에 반영
